third_party/blink/web_tests/svg/dom/no-value-synching-after-property-get.html

<!DOCTYPE html>
<title>Getting a property should not update the corresponding attribute</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg>
  <rect width="100" height="50" fill="green" transform="scale(1)"/>
</svg>
<script>
test(function() {
  var rect = document.querySelector('rect');
  var clone = rect.cloneNode(true);

  clone.systemLanguage;
  clone.transform.baseVal.getItem(0).setTranslate(0, 50);

  rect.parentNode.appendChild(clone);
  assert_false(clone.hasAttribute('systemLanguage'));
});
</script>
